# backend/docs/swagger.yaml
basePath: /
definitions:
...
  internal_routes.GetUsernameByIdRequest:
    properties:
      user_id:
        type: integer
    type: object
  internal_routes.GetUsernameByIdResponse:
    properties:
      username:
        type: string
    type: object
  internal_routes.RefreshTokensRequest:
    properties:
      refresh_token:
        type: string
      userAgent:
        type: string
    type: object
  internal_routes.Response:
    properties:
      access_token:
        type: string
      error:
        type: string
      refresh_token:
        type: string
      status:
        type: string
    type: object
...
  internal_routes.SignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
      userAgent:
        type: string
      username:
        type: string
    type: object
  internal_routes.SignOutRequest:
    properties:
      refresh_token:
        type: string
    type: object
  internal_routes.SignUpRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:8082
info:
  contact: {}
  description: API server for stream service application
  title: Stream Service Prototype API
  version: "1.0"
paths:
  /api/auth/refresh-tokens:
    post:
      consumes:
      - application/json
      description: refresh access & refresh tokens
      operationId: refresh-tokens
      parameters:
      - description: refresh
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_routes.RefreshTokensRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_routes.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_routes.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_routes.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_routes.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/internal_routes.Response'
      summary: RefreshTokens
      tags:
      - Auth
  /api/auth/sign-in:
    post:
      consumes:
      - application/json
      description: Авторизация пользователя с возвратом access и refresh токенов
      operationId: sign-in-user
      parameters:
      - description: Информация для авторизации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_routes.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная авторизация
          schema:
            $ref: '#/definitions/internal_routes.Response'
        "400":
          description: Ошибка запроса
          schema:
            $ref: '#/definitions/internal_routes.Response'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/internal_routes.Response'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/internal_routes.Response'
      summary: Sign In
      tags:
      - Auth
  /api/auth/sign-out:
    post:
      consumes:
      - application/json
      description: delete refresh token
      operationId: delete-refresh-tokens
      parameters:
      - description: Удаляет рефреш токен из таблицы refresh_tokens
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_routes.SignOutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_routes.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_routes.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_routes.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_routes.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/internal_routes.Response'
      summary: SignOut
      tags:
      - Auth
  /api/auth/sign-up:
    post:
      consumes:
      - application/json
      description: create account
      operationId: create-account
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_routes.SignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_routes.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_routes.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_routes.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_routes.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/internal_routes.Response'
      summary: signUp
      tags:
      - Auth
...
  /api/user/get-username-by-id:
    post:
      consumes:
      - application/json
      description: get username by user id
      operationId: get-username-by-id
      parameters:
      - description: Возвращает имя пользователя по user id
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_routes.GetUsernameByIdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_routes.GetUsernameByIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_routes.GetUsernameByIdResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_routes.GetUsernameByIdResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_routes.GetUsernameByIdResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/internal_routes.GetUsernameByIdResponse'
      summary: getUsernameById
      tags:
      - User
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"